+++ examples/create_workflow_submission.py
@@ -1,10 +1,13 @@
+import os
 from datetime import datetime
 
 from python_anvil.api import Anvil
 from python_anvil.api_resources.payload import ForgeSubmitPayload
 
 
-API_KEY = 'my-api-key'
+API_KEY = os.environ.get("ANVIL_API_KEY")
+# or set your own key here
+# API_KEY = 'my-api-key'
 
 
 def main():